HOTFIX BUGFIX, do not copy bad range
If some part of the prefix and/or input recognition does not work
AND the candidate looks suspicious, it tries to copy negative number
of bytes, which is retyped to type_t, thus a large number; leading
to access to bad parts of the memmory and segfault.

Now it checks, if the number of bytes to be copied is zero or positive.